backend/pyproject.toml
[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sarthiai-backend"
version = "0.1.0"
description = "Sarthi AI FastAPI service"
authors = [{ name = "Sarthi AI Team" }]
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.29",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "opik>=1.9.0",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.25",
]

[project.scripts]
sarthiai-scheduler = "app.worker.scheduler_main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*"]

[tool.pytest.ini_options]
pythonpath = ["app"]
addopts = "-q"
